M2k: Implement mechanism to store and retrieve calibration data
Rational:
We need a mechanism to store and retrieve calibration values from
the ADALM-2000. We further need to allow the end user to update these
calibration values without replacing or deleting the initial
factory calibration values. In order that applications like Scopy
can use these values they need to be available via the IIO context.
Therefore we’ve chosen to provide them as context attributes.

The calibration file consist of an arbitrary number of cal,key=value pairs.
For them to become context attributes they must be prefixed with “cal,”.

Calibration files are stored on the ADALM-2000 SPI NOR flash inside a
JFFS2 partition created on mtd2 and mounted to /mnt_jffs2.
The initial factory calibration file is called m2k-calib-factory.ini
and can only be written once to the USB Mass Storage Drive.
It’s then copied to /mnt_jffs2 – further updates are not possible by
the MSD.

When the system boots it checks /mnt_jffs2 for the presence of
m2k-calib.ini (the user calibration file),
if it’s not present the factory calibration file named
m2k-calib-factory.ini is used instead.

The chosen file is then parsed, IIO context attributes are created
and the file is copied to the USB Mass Storage Drive.
It’s filename on the MSD is always: m2k-calib.ini.
In case the user makes modifications to this file, when the drive is
ejected, it’s then written to the JFFS2 partition as user calibration file.
In case the user deletes it completely, the factory calibration file is
restored as m2k-calib.ini.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>