microblaze_adi_defconfig: include ADI kernel, ethtool, phy-tool & mii-diag
This change is a combination of changes.
It adds support for building the ADI kernel (from master), and several
tools for networking.

Make note that the `master` name has been added simply to support shallow
cloning. If the master is updated in the Github repo, the linux-master
tarball needs to be deleted so that it's re-downloaded.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>