Revert "Update for 2021.02.7"
This rolls back to buildroot tag 2021.02. The process done was:

```
revert=$(git --no-pager log --oneline 2021.02..2021.02.7 | cut -d " "
-f1)
for r in ${revert}; do git revert --no-edit ${r}; done
```

Note that this is a squash of all those reverts. The reason for this is
that it will make the merge against 2022.02.3 much more reliable because
we have a linear history from 2021.02 to 2022.02.3. And we will get back
the things we are reverting anyways (just in a away that won't make git
confused or generate conflicts).