Revert "Update for 2022.02.3"
This rolls back to buildroot tag 2022.02. The process done was:

```
revert=$(git --no-pager log --oneline 2022.02..2022.02.3 | cut -d " " -f1)
for r in ${revert}; do git revert --no-edit ${r}; done
```

Note that this is a squash of all those reverts. The reason for this is
that it will make the merge against 2023.02.5 much more reliable because
we have a linear history from 2022.02 to 2023.02.5. And we will get back
the things we are reverting anyways (just in a away that won't make git
confused or generate conflicts).

Signed-off-by: Nuno Sa <nuno.sa@analog.com>